+++ CHANGELOG.md
@@ -15,7 +15,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 -->
 
-
 ## Unreleased
 ### Added
 - Display revision deadlines and make them editable: [#367](https://github.com/CCSDForge/episciences/issues/367) 
@@ -31,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Integration of api and application for the bibliographical references
 
 ### Changed
+-  repositories config. is now stored in DB. @see /scr/mysql/alter_metadata_sources.sql
 - Answer revision request: [#313](https://github.com/CCSDForge/episciences/issues/313):
     * optional file attachment in "Contact without sending a new version" and "answer without any modifications".
 - UX: icon to make it easier to identify the user in revision requests section